[refactor] move config printing stuff in utils.py
